ai-api-integration/requirements.txt
groq
requests
google-generativeai
cohere